requirements.txt
Flask
connexion
injector==0.12.0
Flask-Injector
fastavro
numpy
pandas
tensorflow
pillow
lxml
matplotlib